Initial revamping of blur verification, now performs the check frame by frame. Tests pending and we need to change the alert to the icon and message notifications
