+++ Azul/Azul/PhotoCaptureDelegate.swift
@@ -10,11 +10,6 @@ import Photos
 import Accelerate
 
 class PhotoCaptureProcessor: NSObject {
-    let laplacian: [Float] = [-1, -1, -1,
-                              -1,  8, -1,
-                              -1, -1, -1]
-    
-    var blurryPhoto: Bool = false;
     
     private(set) var requestedPhotoSettings: AVCapturePhotoSettings
     
@@ -25,8 +20,6 @@ class PhotoCaptureProcessor: NSObject {
     lazy var context = CIContext()
     
     private let completionHandler: (PhotoCaptureProcessor) -> Void
-    
-    private let errorHandler: (PhotoCaptureProcessor) -> Void
     
     private let photoProcessingHandler: (Bool) -> Void
     
@@ -44,18 +37,16 @@ class PhotoCaptureProcessor: NSObject {
          willCapturePhotoAnimation: @escaping () -> Void,
          livePhotoCaptureHandler: @escaping (Bool) -> Void,
          completionHandler: @escaping (PhotoCaptureProcessor) -> Void,
-         photoProcessingHandler: @escaping (Bool) -> Void,
-         errorHandler: @escaping (PhotoCaptureProcessor) -> Void) {
+         photoProcessingHandler: @escaping (Bool) -> Void) {
         self.requestedPhotoSettings = requestedPhotoSettings
         self.willCapturePhotoAnimation = willCapturePhotoAnimation
         self.livePhotoCaptureHandler = livePhotoCaptureHandler
         self.completionHandler = completionHandler
         self.photoProcessingHandler = photoProcessingHandler
-        self.errorHandler = errorHandler;
     }
     
     private func error() {
-        errorHandler(self)
+        fatalError("Something happened to the photo while capturing.");
     }
     
     private func didFinish() {
@@ -146,99 +137,6 @@ extension PhotoCaptureProcessor: AVCapturePhotoCaptureDelegate {
         semanticSegmentationMatteDataArray.append(imageData)
     }
     
-    func getLaplacianStDev(data: UnsafeMutableRawPointer,
-                           rowBytes: Int,
-                           width: Int, height: Int,
-                           sequenceCount: Int,
-                           expectedCount: Int,
-                           orientation: UInt32? ) -> Float {
-        var sourceBuffer = vImage_Buffer(data: data,
-                                         height: vImagePixelCount(height),
-                                         width: vImagePixelCount(width),
-                                         rowBytes: rowBytes)
-        
-        var floatPixels: [Float]
-        let count = width * height
-        
-        if sourceBuffer.rowBytes == width * MemoryLayout<Pixel_8>.stride {
-            let start = sourceBuffer.data.assumingMemoryBound(to: Pixel_8.self)
-            floatPixels = vDSP.integerToFloatingPoint(
-                UnsafeMutableBufferPointer(start: start,
-                                           count: count),
-                floatingPointType: Float.self)
-        } else {
-            floatPixels = [Float](unsafeUninitializedCapacity: count) {
-                buffer, initializedCount in
-                
-                var floatBuffer = vImage_Buffer(data: buffer.baseAddress,
-                                                height: sourceBuffer.height,
-                                                width: sourceBuffer.width,
-                                                rowBytes: width * MemoryLayout<Float>.size)
-                
-                vImageConvert_Planar8toPlanarF(&sourceBuffer,
-                                               &floatBuffer,
-                                               0, 255,
-                                               vImage_Flags(kvImageNoFlags))
-                
-                initializedCount = count
-            }
-        }
-        
-        // Convolve with Laplacian.
-        vDSP.convolve(floatPixels,
-                      rowCount: height,
-                      columnCount: width,
-                      with3x3Kernel: laplacian,
-                      result: &floatPixels)
-        
-        // Calculate standard deviation.
-        var mean = Float.nan
-        var stdDev = Float.nan
-        
-        vDSP_normalize(floatPixels, 1,
-                       nil, 1,
-                       &mean, &stdDev,
-                       vDSP_Length(count))
-        return stdDev
-    }
-    
-    func photoIsBlurry(photo: AVCapturePhoto) -> Bool {
-        // Copy the pixel buffer to use it for our blur detection.
-        guard let pixelBuffer = photo.pixelBuffer else {
-            fatalError("Error acquiring pixel buffer.")
-        }
-        
-        CVPixelBufferLockBaseAddress(pixelBuffer,
-                                     CVPixelBufferLockFlags.readOnly)
-        
-        let width = CVPixelBufferGetWidthOfPlane(pixelBuffer, 0)
-        let height = CVPixelBufferGetHeightOfPlane(pixelBuffer, 0)
-        let count = width * height
-        
-        let lumaBaseAddress = CVPixelBufferGetBaseAddressOfPlane(pixelBuffer, 0)
-        let lumaRowBytes = CVPixelBufferGetBytesPerRowOfPlane(pixelBuffer, 0)
-        
-        let lumaCopy = UnsafeMutableRawPointer.allocate(byteCount: count,
-                                                        alignment: MemoryLayout<Pixel_8>.alignment)
-        lumaCopy.copyMemory(from: lumaBaseAddress!,
-                            byteCount: count)
-        
-        
-        CVPixelBufferUnlockBaseAddress(pixelBuffer,
-                                       CVPixelBufferLockFlags.readOnly)
-        
-        let stDev = self.getLaplacianStDev(data: lumaCopy,
-                                  rowBytes: lumaRowBytes,
-                                  width: width,
-                                  height: height,
-                                  sequenceCount: photo.sequenceCount,
-                                  expectedCount: photo.resolvedSettings.expectedPhotoCount,
-                                  orientation: photo.metadata[ String(kCGImagePropertyOrientation) ] as? UInt32)
-        lumaCopy.deallocate()
-        
-        return stDev < 20;
-    }
-    
     /// - Tag: DidFinishProcessingPhoto
     func photoOutput(_ output: AVCapturePhotoOutput, didFinishProcessingPhoto photo: AVCapturePhoto, error: Error?) {
         photoProcessingHandler(false)
@@ -247,10 +145,6 @@ extension PhotoCaptureProcessor: AVCapturePhotoCaptureDelegate {
             print("Error capturing photo: \(error)")
         } else {
             photoData = photo.fileDataRepresentation()
-        }
-        
-        if photoIsBlurry(photo: photo) {
-            self.blurryPhoto = true
         }
         
         // A portrait effects matte gets generated only if AVFoundation detects a face.
@@ -306,54 +200,47 @@ extension PhotoCaptureProcessor: AVCapturePhotoCaptureDelegate {
             return
         }
         
-        if blurryPhoto {
-            print("Photo is blurry. Photo was not saved. Please try again.")
-            blurryPhoto = false
-            self.error()
-            return
-        } else {
-            PHPhotoLibrary.requestAuthorization { status in
-                if status == .authorized {
-                    PHPhotoLibrary.shared().performChanges({
-                        let options = PHAssetResourceCreationOptions()
-                        let creationRequest = PHAssetCreationRequest.forAsset()
-                        options.uniformTypeIdentifier = self.requestedPhotoSettings.processedFileType.map { $0.rawValue }
-                        creationRequest.addResource(with: .photo, data: photoData, options: options)
-                        
-                        if let livePhotoCompanionMovieURL = self.livePhotoCompanionMovieURL {
-                            let livePhotoCompanionMovieFileOptions = PHAssetResourceCreationOptions()
-                            livePhotoCompanionMovieFileOptions.shouldMoveFile = true
-                            creationRequest.addResource(with: .pairedVideo,
-                                                        fileURL: livePhotoCompanionMovieURL,
-                                                        options: livePhotoCompanionMovieFileOptions)
-                        }
-                        
-                        // Save Portrait Effects Matte to Photos Library only if it was generated
-                        if let portraitEffectsMatteData = self.portraitEffectsMatteData {
-                            let creationRequest = PHAssetCreationRequest.forAsset()
-                            creationRequest.addResource(with: .photo,
-                                                        data: portraitEffectsMatteData,
-                                                        options: nil)
-                        }
-                        // Save Portrait Effects Matte to Photos Library only if it was generated
-                        for semanticSegmentationMatteData in self.semanticSegmentationMatteDataArray {
-                            let creationRequest = PHAssetCreationRequest.forAsset()
-                            creationRequest.addResource(with: .photo,
-                                                        data: semanticSegmentationMatteData,
-                                                        options: nil)
-                        }
-                        
-                    }, completionHandler: { _, error in
-                        if let error = error {
-                            print("Error occurred while saving photo to photo library: \(error)")
-                        }
-                        
-                        self.didFinish()
+        PHPhotoLibrary.requestAuthorization { status in
+            if status == .authorized {
+                PHPhotoLibrary.shared().performChanges({
+                    let options = PHAssetResourceCreationOptions()
+                    let creationRequest = PHAssetCreationRequest.forAsset()
+                    options.uniformTypeIdentifier = self.requestedPhotoSettings.processedFileType.map { $0.rawValue }
+                    creationRequest.addResource(with: .photo, data: photoData, options: options)
+                    
+                    if let livePhotoCompanionMovieURL = self.livePhotoCompanionMovieURL {
+                        let livePhotoCompanionMovieFileOptions = PHAssetResourceCreationOptions()
+                        livePhotoCompanionMovieFileOptions.shouldMoveFile = true
+                        creationRequest.addResource(with: .pairedVideo,
+                                                    fileURL: livePhotoCompanionMovieURL,
+                                                    options: livePhotoCompanionMovieFileOptions)
                     }
-                    )
-                } else {
+                    
+                    // Save Portrait Effects Matte to Photos Library only if it was generated
+                    if let portraitEffectsMatteData = self.portraitEffectsMatteData {
+                        let creationRequest = PHAssetCreationRequest.forAsset()
+                        creationRequest.addResource(with: .photo,
+                                                    data: portraitEffectsMatteData,
+                                                    options: nil)
+                    }
+                    // Save Portrait Effects Matte to Photos Library only if it was generated
+                    for semanticSegmentationMatteData in self.semanticSegmentationMatteDataArray {
+                        let creationRequest = PHAssetCreationRequest.forAsset()
+                        creationRequest.addResource(with: .photo,
+                                                    data: semanticSegmentationMatteData,
+                                                    options: nil)
+                    }
+                    
+                }, completionHandler: { _, error in
+                    if let error = error {
+                        print("Error occurred while saving photo to photo library: \(error)")
+                    }
+                    
                     self.didFinish()
                 }
+                )
+            } else {
+                self.didFinish()
             }
         }
     }
